меняет имена компонентов
